+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release History
 
+# 1.28.0 (TBD)
+
+### Snowpark Python API Updates
+
+#### Deprecations:
+
+- Deprecated Snowpark Python function `snowflake_cortex_summarize`. Users can install snowflake-ml-python and use the snowflake.cortex.summarize function instead.
+- Deprecated Snowpark Python function `snowflake_cortex_sentiment`. Users can install snowflake-ml-python and use the snowflake.cortex.sentiment function instead.
+
+#### New Features
+
+- Added support for the following functions in `functions.py`
+  - `normal`
+  - `randn`
+
+### Snowpark pandas API Updates
+
+#### New Features
+
+- Added support for applying Snowflake Cortex functions `Summarize` and `Sentiment`.
+
 ## 1.27.0 (2025-02-03)
 
 ### Snowpark Python API Updates
@@ -76,6 +97,7 @@
   - `snowflake.core` is a dependency required for this feature.
 - Allow user input schema when reading JSON file on stage.
 - Added support for specifying a schema string (including implicit struct syntax) when calling `DataFrame.create_dataframe`.
+- Added support for writing pyarrow Tables to Snowflake tables.
 
 #### Improvements